refactor(alias): Include straggler modules that were accidently not included in previous commit, to update use of moved alias module.
