+++ fhy/lang/ast/serialization/to_json.py
@@ -26,8 +26,8 @@ from typing import Any, Callable, List, Optional, Sequence, Union
 from fhy import ir
 from fhy.lang import ast
 from fhy.lang.ast import visitor
+from fhy.lang.ast.alias import ASTObject
 from fhy.lang.ast.span import Source, Span
-from fhy.utils.alias import ASTObject
 
 
 def convert(value: object) -> Any:
